requirements.txt
django>=4.2.0,<5
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.3.1,<5.4
drf-spectacular>=0.27.0,<0.28
flake8>=7.0.0,<7.1
Pillow>=10.2.0,<10.3
